rename controller + model files, change imports/exports to match
