+++ app.js
@@ -1,6 +1,6 @@
 const express = require("express");
 
-const { getCategories } = require("./controllers/controller");
+const { getCategories } = require("./controllers/categories.controller");
 
 const app = express();
 app.use(express.json());
