+++ controllers/categories.controller.js
@@ -1,4 +1,4 @@
-const { fetchCategories } = require("../models/model");
+const { fetchCategories } = require("../models/categories.model");
 
 exports.getCategories = (req, res, next) => {
   fetchCategories()
